requirements.txt
streamlit>=1.36
pandas>=2.0
openpyxl>=3.1
Pillow>=10.0
pyzbar>=0.1.9
streamlit-webrtc>=0.47
opencv-python-headless>=4.10
av>=11.0
numpy>=1.26
